Add mutable accessors for some important Vec/HashMap fields
Part of https://github.com/containers/oci-spec-rs/issues/86

An obvious operation to support is generating a derived container
image.  Here we want to take some existing data such as the image
manifest and configuration, and append to key fields such as
the image layers, the config diffids and the history.

Another common operation is adding/changing labels.

Having to create a deep-copy of all this data is unfortunate.
I am not very concerned about performance here, but it's also
an ergonomic hit.

Obviously giving raw `&mut` access to fields means that it goes
"behind the back" of the derive builder, and if we ever started
trying to use more advanced things like validation, that wouldn't
work.

Honestly, I am increasingly convinced the builder thing
is just not worth it and it'd be better to have plain structs with
`pub` fields.  But for now, let's allow direct mutation.