Do not panic if windows.layerFolder is null
If windows layerFolders in null string then loading of the runtime spec fails. This happens when containerd serializes Windows{} portion of the spec due to the way golang serializes lists. Although the runtime spec says this is a required field, containerd 1.6 and 1.7 do not fill this field resulting in a null value when serialized to disk. See https://github.com/opencontainers/runtime-spec/issues/1185 for discusion of this field in the runtime spec.

Signed-off-by: James Sturtevant <jstur@microsoft.com>