image/config: Add helpers for config and annotations (labels)
This is an extension of my previous PR to add a `version` API.

Basically, there are so many `Option<T>` to plow through to access
the labels given an ImageConfiguration.  Plus additionally it's
more ergonomic to interact with `&str` and not `&String` in
cases where auto-deref doesn't apply.

I tried to extend the existing config, but what I hit is that
because we deserialize the labels as `HashMap`, we don't necessarily
preserve the order...which is probably its own bug.

Signed-off-by: Colin Walters <walters@verbum.org>