enhancement: Enhance Enum types with string conversion capabilities
For certain enum types, we need to convert them to their corresponding
strings, or be able to map a given string to the corresponding enum type.
However, many enum types currently lack these helper functions. With the
help with strum macro, we can easily reach this goal. But we also take
some special cases which are partially implemented into account.

Fixes #173

Signed-off-by: Alex Lyn <alex.lyn@antgroup.com>