digest: Add AsRef<str>
Conceptually `Digest()` is just a parsed string wrapper. Many
cases want to get access to the full value without allocating,
and the introduction of the type was a regression from that point of
view in 0.7.

Luckily, we can change our internals in such a way that it's
safe to add the impl. Do that by holding onto the full value,
with a duplicate small boxed str for the algorithm only in
the degenerate case that it's an unknown type.

If we had this it would have made https://github.com/containers/bootc/issues/800
less likely.

Signed-off-by: Colin Walters <walters@verbum.org>